Create the translations in batches
When translating a long list of words in one go, the program creates
an oversized matrix to get the results. Since translations of
individual words are fully independent, we can do the computation in
batches, avoiding the huge multi-gigabyte allocation. This change
decreases memory requirements from ~80 GiB to ~20 GiB on our data.